instrumentation-loongsuite/loongsuite-instrumentation-google-adk/tests/requirements.latest.txt
# Core testing (same versions as oldest, but using OTEL from source where applicable)
pytest==7.4.4
pytest-cov>=4.0.0
pytest-asyncio==0.21.0
pytest-vcr==1.0.2
vcrpy>=5.1.0
pyyaml>=6.0.0

# Third-party libraries used in tests
google-adk>=0.1.0
wrapt==1.17.3

# Mock and utilities
mock>=4.0.0

# Internal / OTEL from source
-e opentelemetry-instrumentation
-e instrumentation-loongsuite/loongsuite-instrumentation-google-adk
-e util/opentelemetry-util-genai
